requirements.txt
# Crypto AI Analyzer Dependencies

# HTTP client for API requests
aiohttp>=3.8.0

# WebSocket client for real-time data
websocket-client>=1.6.0

# Async support
asyncio-mqtt>=0.11.0

# Telegram Bot API
python-telegram-bot>=20.7

# Optional: For more advanced technical analysis
# numpy>=1.21.0
# pandas>=1.3.0

# Optional: For real AI integration (uncomment when ready to use)
# openai>=1.0.0
# anthropic>=0.3.0

# Development and testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0 